+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## [0.68.1] - 2024-06-16
+
+### Added
+
+- Added `App.simulate_key`
+
+### Fixed
+
+- Fixed issue with pop_screen launched from an action
+
 ## [0.68.0] - 2024-06-14
 
 ### Added
@@ -2132,6 +2143,7 @@ https://textual.textualize.io/blog/2022/11/08/version-040/#version-040
 - New handler system for messages that doesn't require inheritance
 - Improved traceback handling
 
+[0.68.1]: https://github.com/Textualize/textual/compare/v0.68.0...v0.68.1
 [0.68.0]: https://github.com/Textualize/textual/compare/v0.67.1...v0.68.0
 [0.67.1]: https://github.com/Textualize/textual/compare/v0.67.0...v0.67.1
 [0.67.0]: https://github.com/Textualize/textual/compare/v0.66.0...v0.67.0